requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
langdetect>=1.0.9
pycountry>=23.12.11
openai>=1.0.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0
pdfplumber>=0.10.0
PyPDF2>=3.0.1
pymupdf>=1.23.0